cleanup in filesys.cpp

+++ src/include/fsdb.h
@@ -75,9 +75,9 @@ typedef struct a_inode_struct {
     uae_u32 uniq;
     /* For a directory that is being ExNext()ed, the number of child ainos
        which must be kept locked in core.  */
-    unsigned long locked_children;
+    unsigned int locked_children;
     /* How many ExNext()s are going on in this directory?  */
-    unsigned long exnext_count;
+    unsigned int exnext_count;
     /* AmigaOS locking bits.  */
     int shlock;
     long db_offset;
